fix: mover constantes de auth E2E a archivo separado para evitar import circular
